Correctly handle big traces (>4GB).

+++ src/trunk/sniffers/sniffer-erf.c
@@ -66,8 +66,8 @@ struct erf_me {
     glob_t		files;		/* result of device pattern */
     size_t		file_idx;	/* index of current file in
 					   files.gl_pathv */
-    size_t		file_size;	/* size fo trace file */
-    size_t		nread;
+    off_t		file_size;	/* size fo trace file */
+    off_t		nread;
     size_t		map_size;	/* size of mmap */
     char *		base;		/* mmap addres */
     off_t		off;
@@ -302,7 +302,7 @@ sniffer_next(sniffer_t * s, int max_pkts, timestamp_t max_ivl,
 	int len;		/* total record length */
 	int l2type;		/* interface type */
 	size_t rs;
-	size_t left = me->file_size - me->nread;
+	off_t left = me->file_size - me->nread;
 	char *base = me->base + me->off;
 
 	if (me->nread > me->remap) {
